update script db and api get, add beer

+++ pkg/domain/entities/beer.go
@@ -3,9 +3,9 @@ package entities
 import "time"
 
 type Beer struct {
-	ID         int64     `json:"id"`
+	ID         int       `json:"id"`
 	BeerName   string    `json:"beer_name"`
-	BeerTypeID int64     `json:"beer_type_id"`
+	BeerTypeID int       `json:"beer_type_id"`
 	Image      string    `json:"image"`
 	CreatedAt  time.Time `json:"created_at"`
 	UpdatedAt  time.Time `json:"updated_at"`
